requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
aiomysql>=0.2.0
pydantic[email]>=2.10.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
redis>=5.2.0
aiosmtplib>=3.0.2
python-multipart>=0.0.18
oss2>=2.19.0
Pillow>=11.0.0
httpx>=0.27.0
